tileserv/pyproject.toml
[tool.poetry]
name = "tileserv"
version = "0.1.0"
description = "pg_tileserv database migrations"
authors = ["Conor Brady <con.brad@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^1.4.44"
alembic = "^1.8.1"
psycopg2-binary = "^2.9.5"
python-decouple = "^3.6"
geoalchemy2 = "^0.12.5"
shapely = "^1.8.5.post1"
pytz = "^2022.6"
pytest = "^7.2.0"
aiobotocore = "^2.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
